Move the 'allow_origin' to an attribute
The plugin-object is instantiated only once and not for every use of the
keyword, so we need to do extra bookkeeping for things related to
different endpoints. The new allow_origin attribute is a HashRef that
will keep the 'cors_allow_origin' per $endpoint, so we can have diffent
allowed origins for different endpoints.